requirements.txt
feedparser
requests
requests-oauthlib
python-dotenv
